moved NexusSSOLogin to nxmaccessmanager added logs for manual key validation, will need to rewrite a few things
